+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for bearer token authentication (#40)
+
 ### Changed
 - mujmap now prints a more comprehensive guide on how to recover from a missing
   state file. (#15)